rename forceSourceDeploy to baseDeployCommand
